adk_agents/server_setup.py
import uvicorn
from google.adk.agents import Agent
from tauri_build_agent.agent import root_agent as tauri_agent
from frontend_standard_agent.agent import root_agent as frontend_agent
from fastapi import FastAPI

# Criar apps FastAPI wrappers para cada agente
# Nota: Em um cenário real ADK, isso seria feito automaticamente pelo framework.
# Aqui estou simulando para levantar o servidor.

# Como o ADK funciona: geralmente o 'Agent' já tem um método .app() ou similar
# ou usamos o ModelContainer.

# Vamos assumir que podemos rodar o agente diretamente se ele for compatível com uvicorn
# Se não, vamos usar o padrão do ADK para servir.

# Verificando documentação implícita: O ADK geralmente expõe o app via framework.
# Vou tentar criar um script simples que importa o agente e cria o app.
