adk_agents/requirements.txt
google-adk>=0.1.0
uvicorn
fastapi
